Move view decorators to `site-v2` module
Decorators really apply to the site more than to the CaeNDR package, and they'll want to use some utils from `site-v2/base`.
