+++ src/modules/site-v2/base/utils/view_decorators.py
